+++ spec/rubocop/cop/rails/date_spec.rb
@@ -2,7 +2,8 @@
 
 RSpec.describe RuboCop::Cop::Rails::Date, :config do
   context 'when EnforcedStyle is "strict"' do
-    let(:cop_config) { { 'EnforcedStyle' => 'strict' } }
+    let(:cop_config) { { 'EnforcedStyle' => 'strict', 'AllowToTime' => allow_to_time } }
+    let(:allow_to_time) { true }
 
     shared_examples 'offense' do |method, message|
       it "registers an offense for #{method}" do
@@ -47,25 +48,29 @@ RSpec.describe RuboCop::Cop::Rails::Date, :config do
       end
     end
 
-    %w[to_time to_time_in_current_zone].each do |method|
-      it "registers an offense for ##{method}" do
-        offenses = inspect_source("date.#{method}")
-        expect(offenses.size).to eq(1)
+    context 'when using `to_time_in_current_zone`' do
+      it 'registers an offense' do
+        expect_offense(<<~RUBY)
+          date.to_time_in_current_zone
+               ^^^^^^^^^^^^^^^^^^^^^^^ `to_time_in_current_zone` is deprecated. Use `in_time_zone` instead.
+        RUBY
       end
 
       context 'when using safe navigation operator' do
-        it "registers an offense for ##{method}" do
-          offenses = inspect_source("date&.#{method}")
-          expect(offenses.size).to eq(1)
+        it 'registers an offense' do
+          expect_offense(<<~RUBY)
+            date&.to_time_in_current_zone
+                  ^^^^^^^^^^^^^^^^^^^^^^^ `to_time_in_current_zone` is deprecated. Use `in_time_zone` instead.
+          RUBY
         end
       end
 
-      it "accepts variable named #{method}" do
-        expect_no_offenses("#{method} = 1")
+      it 'accepts variable named `to_time_in_current_zone`' do
+        expect_no_offenses('to_time_in_current_zone = 1')
       end
 
-      it "accepts variable #{method} as range end" do
-        expect_no_offenses("date..#{method}")
+      it 'accepts variable `to_time_in_current_zone` as range end' do
+        expect_no_offenses('date..to_time_in_current_zone')
       end
     end
 
@@ -81,12 +86,91 @@ RSpec.describe RuboCop::Cop::Rails::Date, :config do
       end
     end
 
-    context 'when a string literal without timezone' do
-      it 'registers an offense' do
-        expect_offense(<<~RUBY)
-          "2016-07-12 14:36:31".to_time(:utc)
-                                ^^^^^^^ Do not use `to_time` on Date objects, because they know nothing about the time zone in use.
-        RUBY
+    context 'when using `to_time`' do
+      context 'when `AllowToTime: true`' do
+        let(:allow_to_time) { true }
+
+        it 'does not register an offense' do
+          expect_no_offenses(<<~RUBY)
+            date.to_time
+          RUBY
+        end
+
+        context 'when using safe navigation operator' do
+          it 'does not register an offense' do
+            expect_no_offenses(<<~RUBY)
+              date&.to_time
+            RUBY
+          end
+        end
+
+        it 'accepts variable named `to_time`' do
+          expect_no_offenses('to_time = 1')
+        end
+
+        it 'accepts variable `to_time` as range end' do
+          expect_no_offenses('date..to_time')
+        end
+
+        context 'when a string literal without timezone' do
+          it 'does not register an offense' do
+            expect_no_offenses(<<~RUBY)
+              "2016-07-12 14:36:31".to_time(:utc)
+            RUBY
+          end
+        end
+
+        RuboCop::Cop::Rails::TimeZone::ACCEPTED_METHODS.each do |a_method|
+          it "does not register an offense for val.to_time.#{a_method}" do
+            expect_no_offenses(<<~RUBY)
+              val.to_time.#{a_method}
+            RUBY
+          end
+        end
+      end
+
+      context 'when `AllowToTime: false`' do
+        let(:allow_to_time) { false }
+
+        it 'registers an offense' do
+          expect_offense(<<~RUBY)
+            date.to_time
+                 ^^^^^^^ Do not use `to_time` on Date objects, because they know nothing about the time zone in use.
+          RUBY
+        end
+
+        context 'when using safe navigation operator' do
+          it 'registers an offense' do
+            expect_offense(<<~RUBY)
+              date&.to_time
+                    ^^^^^^^ Do not use `to_time` on Date objects, because they know nothing about the time zone in use.
+            RUBY
+          end
+        end
+
+        it 'accepts variable named `to_time`' do
+          expect_no_offenses('to_time = 1')
+        end
+
+        it 'accepts variable `to_time` as range end' do
+          expect_no_offenses('date..to_time')
+        end
+
+        context 'when a string literal without timezone' do
+          it 'registers an offense' do
+            expect_offense(<<~RUBY)
+              "2016-07-12 14:36:31".to_time(:utc)
+                                    ^^^^^^^ Do not use `to_time` on Date objects, because they know nothing about the time zone in use.
+            RUBY
+          end
+        end
+
+        RuboCop::Cop::Rails::TimeZone::ACCEPTED_METHODS.each do |a_method|
+          it "registers an offense for val.to_time.#{a_method}" do
+            offenses = inspect_source("val.to_time.#{a_method}")
+            expect(offenses.size).to eq(1)
+          end
+        end
       end
     end
 
@@ -98,13 +182,6 @@ RSpec.describe RuboCop::Cop::Rails::Date, :config do
 
     it 'does not blow up in the presence of a single constant to inspect' do
       expect_no_offenses('A')
-    end
-
-    RuboCop::Cop::Rails::TimeZone::ACCEPTED_METHODS.each do |a_method|
-      it "registers an offense for val.to_time.#{a_method}" do
-        offenses = inspect_source("val.to_time.#{a_method}")
-        expect(offenses.size).to eq(1)
-      end
     end
 
     it 'registers an offense for #to_time_in_current_zone' do