+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bug fixes
 
+* [#482](https://github.com/rubocop/rubocop-rails/pull/482): Fix a false positive for `Rails/RelativeDateConstant` when assigning (hashes/arrays/etc)-containing procs to a constant. ([@jdelStrother][])
 * [#419](https://github.com/rubocop/rubocop-rails/issues/419): Fix an error for `Rails/UniqueValidationWithoutIndex` when using a unique index and `check_constraint` that has `nil` first argument. ([@koic][])
 
 ## 2.10.1 (2021-05-06)
@@ -382,3 +383,4 @@
 [@leonp1991]: https://github.com/leonp1991
 [@drenmi]: https://github.com/drenmi
 [@rhymes]: https://github.com/rhymes
+[@jdelStrother]: https://github.com/jdelStrother