+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ### Changes
 
 * [#288](https://github.com/rubocop/rubocop-rails/issues/288): Add `AllowToTime` option (`true` by default) to `Rails/Date`. ([@koic][])
+* [#499](https://github.com/rubocop/rubocop-rails/issues/499): Add `IgnoreWhereFirst` option (`true` by default) to `Rails/FindBy`. ([@koic][])
 
 ## 2.10.1 (2021-05-06)